refactor: move flattener.py under oracle/
