+++ tenforty/oracle/flattener.py
